EIP-9 Divided crud_operitions into 2 files to better organize the code
